requirements.txt
python-telegram-bot==20.3
pydub
SpeechRecognition
python-dotenv
ffmpeg-python